requirements.txt
pymongo==4.2.0
PyYAML==6.0